Refactor duplicated mushroom code and take bonemeal when activated.
